# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: OTUS Highload Architect
  version: 1.2.0
servers:
- url: /
paths:
...
  /post/create:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_post_create_post_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostId'
          description: Успешно создан пост
        "400":
          description: Невалидные данные ввода
        "401":
          description: Неавторизованный доступ
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
      security:
      - bearerAuth: []
      x-content-type: application/json
      x-accepts:
      - application/json
  /post/update:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_post_update_put_request'
      responses:
        "200":
          description: Успешно изменен пост
        "400":
          description: Невалидные данные ввода
        "401":
          description: Неавторизованный доступ
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
      security:
      - bearerAuth: []
      x-content-type: application/json
      x-accepts:
      - application/json
  /post/delete/{id}:
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/PostId'
        style: simple
      responses:
        "200":
          description: Успешно удален пост
        "400":
          description: Невалидные данные ввода
        "401":
          description: Неавторизованный доступ
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
      security:
      - bearerAuth: []
      x-accepts:
      - application/json
  /post/get/{id}:
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/PostId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
          description: Успешно получен пост
        "400":
          description: Невалидные данные ввода
        "401":
          description: Неавторизованный доступ
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
      x-accepts:
      - application/json
  /post/feed:
    get:
      parameters:
      - explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          description: Оффсет с которого начинать выдачу
          example: 100
          minimum: 0
          type: number
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          description: "Лимит, ограничивающий кол-во возвращенных сущностей"
          example: 10
          minimum: 1
          type: number
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Post'
                type: array
          description: Успешно получены посты друзей
        "400":
          description: Невалидные данные ввода
        "401":
          description: Неавторизованный доступ
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
      security:
      - bearerAuth: []
      x-accepts:
      - application/json
...
components:
...
  schemas:
...
    PostId:
      description: Идентификатор поста
      example: 1d535fd6-7521-4cb1-aa6d-031be7123c4d
      type: string
...
    Post:
      description: Пост пользователя
      example:
        author_user_id: author_user_id
        id: 1d535fd6-7521-4cb1-aa6d-031be7123c4d
        text: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod\
          \ tempor incididunt ut labore et dolore magna aliqua. Lectus mauris ultrices\
          \ eros in cursus turpis massa."
      properties:
        id:
          description: Идентификатор поста
          example: 1d535fd6-7521-4cb1-aa6d-031be7123c4d
          type: string
        text:
          description: Текст поста
          example: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do\
            \ eiusmod tempor incididunt ut labore et dolore magna aliqua. Lectus mauris\
            \ ultrices eros in cursus turpis massa."
          type: string
        author_user_id:
          description: Идентификатор пользователя
          type: string
      type: object
...
    _login_post_500_response:
      example:
        code: 0
        message: message
        request_id: request_id
      properties:
        message:
          description: Описание ошибки
          type: string
        request_id:
          description: Идентификатор запроса. Предназначен для более быстрого поиска
            проблем.
          type: string
        code:
          description: Код ошибки. Предназначен для классификации проблем и более
            быстрого решения проблем.
          type: integer
      required:
      - message
      type: object
...
    _post_create_post_request:
      properties:
        text:
          description: Текст поста
          example: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do\
            \ eiusmod tempor incididunt ut labore et dolore magna aliqua. Lectus mauris\
            \ ultrices eros in cursus turpis massa."
          type: string
      required:
      - text
      type: object
    _post_update_put_request:
      properties:
        id:
          description: Идентификатор поста
          example: 1d535fd6-7521-4cb1-aa6d-031be7123c4d
          type: string
        text:
          description: Текст поста
          example: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do\
            \ eiusmod tempor incididunt ut labore et dolore magna aliqua. Lectus mauris\
            \ ultrices eros in cursus turpis massa."
          type: string
      required:
      - id
      - text
      type: object
...
  securitySchemes:
    bearerAuth:
      description: "Авторизация по токену, который был получен в методе /login"
      scheme: bearer
      type: http